CHEC-325 ; FPX login page issue fix
chore: Lumberjack: updated the block to maintain/execute/gather the data beforehand due to a compiler error

bug-fix: Removed webview.reload function in webview:didFail function as this wasn't leading to errors. Updated it in the
didFailProvisionalNavigation function as the error was being generated there.

chore: commented print function in all classes